pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "family-agent-os"
version = "0.1.0"
description = "Local-first Family Agent OS runtime scaffold"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "PyYAML>=6.0",
  "python-telegram-bot>=20.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "skills*"]
